refactor: Move stores in @/stores folder
