+++ src/stores/useMeterSettingsStore.ts
